feat: implement solar car AI chatbot interface with interactive components and actions
