+++ src/app/_components/ChatBot/ChatBot.module.scss
@@ -12,7 +12,7 @@
   justify-content: center;
   z-index: 50;
   box-shadow: 0 8px 32px rgba(0, 0, 0, 0.4);
-  background: linear-gradient(to bottom right, #FF8080, #FF0000, #8B0000);
+  background: linear-gradient(to bottom right, #ff8080, #ff0000, #8b0000);
 }
 
 .triggerRing {
@@ -27,7 +27,8 @@
   width: 90%;
   height: auto;
   object-fit: contain;
-  filter: drop-shadow(0 10px 8px rgba(0, 0, 0, 0.04)) drop-shadow(0 4px 3px rgba(0, 0, 0, 0.1));
+  filter: drop-shadow(0 10px 8px rgba(0, 0, 0, 0.04))
+    drop-shadow(0 4px 3px rgba(0, 0, 0, 0.1));
 }
 
 .chatPanel {
@@ -46,7 +47,10 @@
   border: 1px solid rgba(255, 0, 0, 0.2);
   background-color: rgba(8, 16, 26, 0.9);
   backdrop-filter: blur(24px);
-  box-shadow: 0 24px 80px rgba(0, 0, 0, 0.7), 0 0 0 1px rgba(255, 255, 255, 0.05), inset 0 1px 0 rgba(255, 255, 255, 0.08);
+  box-shadow:
+    0 24px 80px rgba(0, 0, 0, 0.7),
+    0 0 0 1px rgba(255, 255, 255, 0.05),
+    inset 0 1px 0 rgba(255, 255, 255, 0.08);
 }
 
 .header {
@@ -65,7 +69,7 @@
 .headerTitle {
   font-weight: 700;
   font-size: 16px;
-  color: #F0EDE8;
+  color: #f0ede8;
   letter-spacing: 0.025em;
 }
 
@@ -78,11 +82,11 @@
 }
 
 .headerStatusLoading {
-  color: #FF0000;
+  color: #ff0000;
 }
 
 .headerStatusReady {
-  color: #4CAF80;
+  color: #4caf80;
 }
 
 .statusDot {
@@ -92,13 +96,13 @@
 }
 
 .statusDotLoading {
-  background-color: #FF0000;
-  box-shadow: 0 0 6px #FF0000;
+  background-color: #ff0000;
+  box-shadow: 0 0 6px #ff0000;
 }
 
 .statusDotReady {
-  background-color: #4CAF80;
-  box-shadow: 0 0 6px #4CAF80;
+  background-color: #4caf80;
+  box-shadow: 0 0 6px #4caf80;
 }
 
 .closeButton {
@@ -116,9 +120,10 @@
   transition: background-color 0.2s;
 }
 
-.closeButton:hover, .closeButton:focus-visible {
+.closeButton:hover,
+.closeButton:focus-visible {
   background-color: rgba(255, 255, 255, 0.1);
-  outline: 2px solid #FF0000;
+  outline: 2px solid #ff0000;
 }
 
 .messagesContainer {
@@ -164,12 +169,12 @@
   font-size: 14.5px;
   line-height: 1.625;
   letter-spacing: -0.015em;
-  color: #F0EDE8;
+  color: #f0ede8;
 }
 
 .messageBubbleUser {
   border-radius: 20px 20px 4px 20px;
-  background: linear-gradient(to bottom right, #FF0000, #8B0000);
+  background: linear-gradient(to bottom right, #ff0000, #8b0000);
   border: none;
   box-shadow: 0 4px 20px rgba(255, 0, 0, 0.3);
 }
@@ -195,15 +200,16 @@
   border-radius: 9999px;
   border: 1px solid rgba(255, 90, 90, 0.25);
   background-color: rgba(255, 90, 90, 0.05);
-  color: #FF5A5A;
+  color: #ff5a5a;
   cursor: pointer;
   transition: all 0.2s;
 }
 
-.suggestionButton:hover, .suggestionButton:focus-visible {
+.suggestionButton:hover,
+.suggestionButton:focus-visible {
   background-color: rgba(255, 90, 90, 0.15);
   border-color: rgba(255, 90, 90, 0.5);
-  outline: 2px solid #FF5A5A;
+  outline: 2px solid #ff5a5a;
 }
 
 .inputForm {
@@ -220,7 +226,7 @@
   border: 1px solid rgba(255, 255, 255, 0.1);
   border-radius: 12px;
   padding: 10px 14px;
-  color: #F0EDE8;
+  color: #f0ede8;
   font-size: 14px;
   resize: none;
   line-height: 1.625;
@@ -246,7 +252,7 @@
 }
 
 .sendButtonActive {
-  background: linear-gradient(to bottom right, #FF0000, #8B0000);
+  background: linear-gradient(to bottom right, #ff0000, #8b0000);
   cursor: pointer;
 }
 
@@ -255,7 +261,7 @@
 }
 
 .sendButtonActive:focus-visible {
-  outline: 2px solid #FF0000;
+  outline: 2px solid #ff0000;
 }
 
 .sendButtonDisabled {
@@ -275,7 +281,7 @@
   width: 7px;
   height: 7px;
   border-radius: 9999px;
-  background-color: #FF0000;
+  background-color: #ff0000;
   opacity: 0.8;
 }
 
@@ -293,6 +299,12 @@
 }
 
 .boldText {
-  color: #FF5A5A;
+  color: #ff5a5a;
   font-weight: bold;
+}
+
+.chatBotContainer {
+  @media (max-width: 767px) {
+    display: none !important;
+  }
 }
