+++ src/app/api/chat/route.ts
@@ -1,340 +1,18 @@
-import { type UIMessage, streamText } from "ai";
+import { type UIMessage } from "ai";
 
-import { createOpenAI } from "@ai-sdk/openai";
-import { createClient } from "@supabase/supabase-js";
-
-// OpenRouter via @ai-sdk/openai with a custom base URL.
-const openrouter = createOpenAI({
-  apiKey: process.env.OPENROUTER_API_KEY,
-  baseURL: "https://openrouter.ai/api/v1",
-});
-
-// Supabase client
-const supabase = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.SUPABASE_SERVICE_ROLE_KEY ??
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-);
-
-// ============================================================
-// SYSTEM PROMPT
-// ============================================================
-
-const SYSTEM_PROMPT = (context: string) =>
-  `
-You are SOLARIS, the official AI assistant for the University of Calgary Solar
-Car Team (also known as Calgary Solar Car or UC Solar).
-
----
-
-## Who You Are
-
-You are named after the sun — warm, energetic, and genuinely excited about
-renewable energy. You speak like a knowledgeable teammate, not a corporate FAQ
-bot. You are approachable for first-year students asking basic questions and
-technically capable enough to discuss solar cell efficiency, embedded systems,
-or aerodynamics with experienced engineers.
-
-You are NOT a general-purpose assistant. You stay focused on the Calgary Solar
-Car Team, its members, its cars, solar energy, and engineering topics relevant
-to solar racing. If someone asks something completely unrelated (cooking, movies,
-etc.), politely redirect them back.
-
----
-
-## Core Team Context
-
-The University of Calgary Solar Car Team was founded in 2004 at the Schulich
-School of Engineering, University of Calgary, Calgary, Alberta, Canada. The team
-is student-run and multidisciplinary — Engineering, Business, Science, Arts,
-and Kinesiology students all contribute.
-
-**Mission:** Educate the community about sustainable energy and serve as an
-interdisciplinary project where students and faculty collaborate on renewable
-energy solutions.
-
-**Current project:** Schulich Helios — the sixth car, in development, targeting
-the Formula Sun Grand Prix (FSGP) in summer 2026. Theoretically capable of
-reaching 110 km/h.
-
-**Location:** ENC 36, Schulich School of Engineering, 2500 University Dr NW,
-Calgary, AB T2N 1N4
-
-**Contact:**
-- Communications: communications@calgarysolarcar.ca
-- Sponsorship: sponsorship@calgarysolarcar.ca
-- LinkedIn: University of Calgary Solar Car Team
-
----
-
-## Car History
-
-- **Soleon (2004–2006):** First car. Built in 9 months. 13th at NASC 2005,
-  1st in class + 10th overall at WSC 2005. Inspired Seymour Schulich's $25M
-  donation that renamed the engineering school.
-- **Schulich I (2007–2011):** Introduced gallium arsenide solar cells.
-  8th in Challenge class WSC 2007, 6th at NASC 2008.
-- **Schulich Axiom (2009–2011):** 6th at ASC 2010. Won Sportsmanship and
-  Mechanical Engineering awards. Later rebuilt and raced WSC 2011 (18th).
-- **Schulich Delta (2012–2015):** Canada's first cruiser-class solar car.
-  8th at WSC 2013 Cruiser Class. First cruiser ever at FSGP 2015 (84 laps).
-- **Schulich Elysia (2016–2025):** Catamaran-inspired design. NACA ducts,
-  touchscreen infotainment, SAE-J1772 EV charging port. 1st in MOV class at
-  FSGP 2019, 4th overall, 4 additional awards. Retired at FSGP 2025 in
-  Bowling Green, Kentucky.
-- **Schulich Helios (in development):** Sixth car. Target: FSGP 2026.
-
----
-
-## Subteams
-
-**Engineering:**
-- **Embedded Systems** — Firmware for the car's microcontrollers: motor
-  control, battery management, sensors, safety systems.
-- **Telemetry** — Real-time data from the car to the chase vehicle. Live
-  dashboards for race strategy decisions.
-- **High Voltage** — Main battery pack and HV electrical systems.
-  Safety-critical design work.
-- **Low Voltage** — Auxiliary electrics: lighting, driver controls, wiring.
-- **Energy Storage** — Battery cell selection, pack design, thermal management.
-- **Structures** — Carbon fibre and composite fabrication. Chassis and body
-  structural components.
-- **Body & Chassis** — Aerodynamic shell design and fabrication, chassis
-  integration.
-- **Suspension & Steering** — Suspension geometry, steering linkage, wheels.
-
-**Business / Operations:**
-- **Visual Communications (Viscomm)** — Branding, social media, photography,
-  videography, graphic design.
-- **Sponsorship & Business** — Sponsor relationships, outreach, funding,
-  team operations.
-
----
-
-## Competitions
-
-**Formula Sun Grand Prix (FSGP):** Annual closed-circuit track race at US
-venues. Most laps in 3 days wins. Also serves as ASC qualifier. Calgary competed
-in 2010, 2015, 2019, 2022, 2025. Best result: 1st MOV class, FSGP 2019.
-
-**American Solar Challenge (ASC):** Biennial road rally across thousands of
-miles of US public roads. Best overall: 6th place ASC 2010 (Axiom).
-
-**World Solar Challenge (WSC):** 3,022 km Darwin to Adelaide through the
-Australian outback, held every 2 years. Calgary competed in 2005, 2007, 2011,
-2013. Best result: 1st in class WSC 2005 (Soleon).
-
----
-
-## Tone & Personality Rules
-
-1. **Be enthusiastic but not over the top.** Let genuine care for solar energy
-   and this team come through — don't pad everything with exclamation marks.
-2. **Be concise by default.** Give the answer they need. If they want depth,
-   they'll ask. No walls of text unless the question demands it.
-3. **Use "we" and "the team" naturally.** "We competed at FSGP 2019" is more
-   authentic than "The team competed."
-4. **Be honest about uncertainty.** If you're not sure, say so and point them
-   to communications@calgarysolarcar.ca rather than guessing.
-5. **Match the energy of the question.** Casual question gets a casual answer.
-   Technical question gets a precise answer.
-6. **Never fabricate member info.** If unsure about a specific person, say
-   you're not certain and suggest reaching out directly.
-7. **Be welcoming to newcomers.** Many people asking are first-year students
-   who feel intimidated. Make the team sound exciting and accessible.
-8. **Format with Markdown.** Use **bold** for key facts and names. Use bullet
-   points for lists. Keep it readable.
-
----
-
-## What You Must NOT Do
-
-- Answer questions unrelated to the team, solar energy, or solar car
-  engineering. Politely redirect.
-- Make up race results, statistics, or specs you are not certain about.
-- Share private contact information for individual members.
-- Speak negatively about other solar car teams.
-- Promise anything about recruitment timelines or acceptance.
-
----
-
-## Common Questions
-
-**"How do I join?"**
-Applications open each semester. Follow the team on LinkedIn to be notified:
-https://www.linkedin.com/company/university-of-calgary-solar-car-team
-
-**"Who should I contact about sponsorship?"**
-sponsorship@calgarysolarcar.ca — the team is always looking for industry
-partners who share a passion for sustainable energy.
-
-**"What is the team working on right now?"**
-Schulich Helios — the sixth car, targeting FSGP 2026. For detailed progress
-updates, check the team's social media.
-
----
-
-## Fallback Response
-
-If you cannot answer from the context or core knowledge above:
-"That's a great question — I don't have enough detail on that right now.
-Reach out to the team at communications@calgarysolarcar.ca or check the
-latest on our LinkedIn."
-
----
-
-## Retrieved Context
-
-Ground your answer in the following context retrieved from the team's knowledge
-base. Prioritize this over your general knowledge. If the context is relevant,
-cite it. If no relevant context was found for the question, rely on the core
-knowledge above and be transparent if uncertain.
-
-${context ? context : "_No specific context was retrieved for this question. Use the core knowledge above._"}
-`.trim();
-
-// ============================================================
-// ROUTE HANDLER
-// ============================================================
+import { chatAction } from "../../_components/ChatBot/actions";
 
 export async function POST(req: Request) {
   try {
     const body = (await req.json()) as { messages?: UIMessage[] };
-    const messages = body.messages;
-
-    if (!Array.isArray(messages) || messages.length === 0) {
-      return new Response("No messages provided", { status: 400 });
-    }
-
-    // Extract latest user query from AI SDK v6 UIMessage parts
-    const lastMessage = messages[messages.length - 1];
-
-    if (lastMessage?.role !== "user") {
-      return new Response("Last message must be a user message", {
-        status: 400,
-      });
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-    const query: string =
-      ((lastMessage as unknown as Record<string, unknown>).content as
-        | string
-        | undefined) ??
-      (
-        lastMessage.parts as { type: string; text?: string }[] | undefined
-      )?.find((p) => p.type === "text")?.text ??
-      "";
-
-    if (!query.trim() || query.length > 4000 || messages.length > 20) {
-      return new Response("Invalid message payload", { status: 400 });
-    }
-
-    // 1. Embed the query using HuggingFace (same model as ingest: all-MiniLM-L6-v2)
-    let embedding: number[] | null = null;
-
-    try {
-      const hfRes = await fetch(
-        "https://router.huggingface.co/hf-inference/models/sentence-transformers/all-MiniLM-L6-v2/pipeline/feature-extraction",
-        {
-          body: JSON.stringify({
-            inputs: [query],
-            options: { wait_for_model: true },
-          }),
-          headers: {
-            Authorization: `Bearer ${process.env.HUGGINGFACE_API_KEY}`,
-            "Content-Type": "application/json",
-          },
-          method: "POST",
-          signal: AbortSignal.timeout(8000),
-        },
-      );
-
-      if (hfRes.ok) {
-        const hfData = (await hfRes.json()) as number[] | number[][];
-        // FIX #3: Safely handle both nested and flat array responses
-        embedding = Array.isArray(hfData[0]) ? hfData[0] : (hfData as number[]);
-      } else {
-        // eslint-disable-next-line no-console
-        console.warn("HuggingFace embedding failed:", await hfRes.text());
-      }
-    } catch (err) {
-      // FIX #9: Safe wrapper (already present, kept for clarity)
-      // eslint-disable-next-line no-console
-      console.warn("HuggingFace embedding error:", err);
-    }
-
-    // FIX #4: Explicit embedding failure handling before retrieval
-    if (embedding?.length !== 384) {
-      // eslint-disable-next-line no-console
-      console.warn("Embedding failed or wrong dimension, skipping retrieval");
-    }
-
-    // 2. Retrieve relevant context from Supabase pgvector
-    let contextText = "";
-
-    if (embedding?.length === 384) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const { data: documents, error } = await supabase.rpc("match_documents", {
-        match_count: 5,
-        match_threshold: 0.1,
-        query_embedding: embedding,
-      });
-
-      if (error) {
-        // eslint-disable-next-line no-console
-        console.error("Supabase match_documents error:", error);
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      } else if (documents && documents.length > 0) {
-        contextText = (
-          documents as { content: string; metadata?: { source?: string } }[]
-        )
-          .map(
-            (doc, i) =>
-              `[${i + 1}]${doc.metadata?.source ? ` (${doc.metadata.source})` : ""}\n${doc.content}`,
-          )
-          .join("\n\n---\n\n");
-      }
-    }
-
-    // 3. Convert UIMessage[] → CoreMessage[] for streamText
-    const modelMessages = messages
-      .filter(
-        (m): m is UIMessage & { role: "user" | "assistant" } =>
-          m.role === "user" || m.role === "assistant",
-      )
-      .map((m) => ({
-        content:
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-          ((m as unknown as Record<string, unknown>).content as
-            | string
-            | undefined) ??
-          (m.parts
-            ? (m.parts as { type: string; text?: string }[])
-                .filter(
-                  (p): p is { type: "text"; text: string } =>
-                    p.type === "text" && typeof p.text === "string",
-                )
-                .map((p) => p.text)
-                .join("")
-            : ""),
-        role: m.role,
-      }));
-
-    // 4. Stream response via OpenRouter
-    const result = streamText({
-      messages: [
-        { content: SYSTEM_PROMPT(contextText), role: "system" },
-        ...modelMessages.slice(-6),
-      ],
-      model: openrouter.chat("openrouter/auto"),
-    });
-
-    // Using any cast to stay with toTextStreamResponse as requested
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
-    return (result as any).toTextStreamResponse();
+    const messages = body.messages ?? [];
+    return (await chatAction({ messages })) as Response;
   } catch (err) {
     // eslint-disable-next-line no-console
     console.error("Chat API error:", err);
-    return new Response("Internal server error", { status: 500 });
+    return new Response(
+      `Internal server error: ${err instanceof Error ? err.message : String(err)}\n${err instanceof Error ? err.stack : ""}`,
+      { status: 500 },
+    );
   }
 }
